move smartify to markup
